+++ src/base/global_guard.rs
