+++ src/base/global_pointer.rs
