+++ src/benchmark/bench_global_guard.rs
@@ -1,11 +1,10 @@
-//use global_guard::GlobalGuard;
-//use config::Config;
+use base::{config::Config, global_guard::GlobalGuard};
+use backend::comm;
+
 use std::time::{SystemTime, Duration};
 use std::vec::Vec;
 //use comm;
 use statistical;
-use backend::comm;
-use central;
 
 fn duration_to_nano(duration: &Duration) -> u128 {
     let in_nanos = duration.as_secs() as u128 * 1000_000_000 +
