+++ src/benchmark/bench_global_pointer.rs
@@ -1,8 +1,8 @@
-use global_pointer::GlobalPointer;
-use config::Config;
+use base::{global_pointer::GlobalPointer, config::Config};
+use backend::comm;
+
 use std::time::{SystemTime, Duration};
 use std::vec::Vec;
-use comm;
 use statistical;
 
 fn duration_to_nano(duration: &Duration) -> u128 {
